src/filterkeys: When filtering on multiple columns use delim as separator.
Filterkeys constructs the filter by concatenating the values of the key
columns. A case came up where the concat of the keys in the filter was
equal to the concat of the keys of the file being filtered but the keys
where different by themselves. Eg:

filter-file:
col1,col2
123,456

file being filtered:
col1,col2
1234,56